+++ src/main/scala/DebugFileReader.scala
@@ -2,7 +2,7 @@ package xyz.hyperreal.m68k
 
 import scala.collection.mutable.HashMap
 
-import java.lang.Long.parseLong
+import java.lang.Integer.parseInt
 
 
 object DebugFileReader extends App {
@@ -13,13 +13,13 @@ object DebugFileReader extends App {
   lazy val varRegex = """^(?:[^ ]| *static).+?([a-zA-Z_][a-zA-Z0-9_]*)[^ ]* /\* 0x([^ ]+) \*/;"""r
 
   def apply( src: io.Source ) = {
-    val code = new HashMap[Long, (String, String)]
-    val vars = new HashMap[Long, String]
+    val code = new HashMap[Int, (String, String)]
+    val vars = new HashMap[Int, String]
 
     for (line <- src.getLines)
       line match {
-        case codeRegex( file, lineno, address ) => code(parseLong( address, 16 )) = (lineno, file)
-        case varRegex( name, address ) => vars(parseLong( address, 16 )) = name
+        case codeRegex( file, lineno, address ) => code(parseInt( address, 16 )) = (lineno, file)
+        case varRegex( name, address ) => vars(parseInt( address, 16 )) = name
         case _ =>
       }
