+++ src/main/scala/MapFileReader.scala
@@ -9,8 +9,8 @@ object MapFileReader extends App {
 
   lazy val labelRegex = " {16}0x([^ ]+) {16}([^. ]+).*"r
 
-  def apply( src: io.Source ): (HashMap[String, Long], HashMap[Long, String]) = {
-    val symbols = new HashMap[String, Long]
+  def apply( src: io.Source ): (HashMap[String, Int], HashMap[Int, String]) = {
+    val symbols = new HashMap[String, Int]
     var started = false
 
     for (line <- src.getLines) {
@@ -20,7 +20,7 @@ object MapFileReader extends App {
             return (symbols, symbols map {case (d, r) => (r, d)})
           else {
             line match {
-              case labelRegex( address, label ) => symbols(label) = java.lang.Long.parseLong( address, 16 )
+              case labelRegex( address, label ) => symbols(label) = java.lang.Integer.parseInt( address, 16 )
               case _ =>
             }
           }
